+++ src/joystick/joystick.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <thread>
 #include <mutex>
+#include <common/utils.h>
 
 #define JOYSTICK_UP_PATH "/sys/class/gpio/gpio26/"
 #define JOYSTICK_DOWN_PATH "/sys/class/gpio/gpio46/"
@@ -29,7 +30,7 @@ class Joystick {
     // instance variables
     std::thread* joystickThread;
     std::thread* decayerThread;
-    Vec2<int> strengthVector;
+    Vec2<double> strengthVector;
     bool isRunning;
     std::mutex strengthMutex;
 
@@ -50,7 +51,7 @@ class Joystick {
 
     void lockStrength(){this->strengthMutex.lock();}
     void unlockStrength(){this->strengthMutex.unlock();}
-    Vec2<int> getStrength(){return this->strengthVector;}
+    Vec2<double> getStrength(){return this->strengthVector;}
     double getStrengthX(){return this->strengthVector.x;}
     double getStrengthZ(){return this->strengthVector.y;}
     void setStrength(double _x, double _z){this->strengthVector.x = _x, this->strengthVector.y = _z;};
